function_app/requirements.txt
azure-functions>=1.17.0,<2.0.0
azure-storage-blob>=12.18.0,<13.0.0
azure-storage-queue>=12.6.0,<13.0.0
azure-search-documents>=11.4.0,<12.0.0
azure-core>=1.30.0,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.0,<2.0.0
asyncpg>=0.27.0,<0.28.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.6
azure-identity>=1.14.0,<2.0.0
opencensus-ext-azure>=1.1.9,<2.0.0
opencensus-ext-logging>=0.1.0,<0.2.0
azure-ai-formrecognizer>=3.3.0,<4.0.0 